Moved to CUDA 11 and Fastvideo SDK 0.17.0.1

+++ src/CameraSample/Camera/XimeaCamera.cpp
@@ -216,7 +216,7 @@ void XimeaCamera::startStreaming()
         return;
 
     QByteArray frameData;
-    frameData.resize(mInputBuffer.size());
+    frameData.resize((int)mInputBuffer.size());
     XI_IMG image = {0};
     image.size = sizeof(XI_IMG);
     image.bp_size = frameData.size();
